+++ simple_file_saver_platform_interface/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.0.0
+
+* BREAKING CHANGE: The platform interface API in now simplified into one single method `saveFile()` and an additional method `downloadLinkBuilder()` only for the web platform.
+* The new API takes a `FileSaveInfo` object which you may create it from either `FileSaveInfo.fromBytes()` or `FileSaveInfo.fromUrl()`.
+
+## 1.3.0
+
+* Return the path of the file saved instead of a boolean.
+
 ## 1.2.0
 
 * Add new platform APIs for the web implementation, and updates the download APIs.